+++ Iris/App/Transport/Device.swift
@@ -13,14 +13,26 @@ class Device {
 	private var endPoint: String!
 
 	func setEndpoint(endpoint: String) {
-		endPoint = endpoint
+        self.endPoint = endpoint
 	}
 	
 	func isReachable() -> Bool {
 		return send(data: [255,255,255])
 	}
+    
+    func sendStaticColor(data: [UInt8]) -> Bool {
+        var container: [UInt8] = [UInt8(2)]
+        container = container + data
+        return send(data: container)
+    }
+    
+    func sendScreenBuffer(data: [UInt8]) -> Bool {
+        var container: [UInt8] = [UInt8(1)]
+        container = container + data
+        return send(data: container)
+    }
 	
-	func send(data: [UInt8]) -> Bool {
+	private func send(data: [UInt8]) -> Bool {
 		do {
 			let socket = try Socket.create(family: Socket.ProtocolFamily.inet)
 			try socket.connect(to: endPoint, port: 8080, timeout: 2000)